refactor(viewer): extract persist config to viewerStore.persistence
Move the migrate function and partialize out of createViewerStore into
viewerStore.persistence.ts, alongside the PersistedViewerState type they
share. partialize is typed `=> PersistedViewerState`, so it and the
migrate fallback stay in sync — drift is a compile error. createViewerStore
is now purely store assembly.
